+++ lib/internal/debugger/_inspect.js
@@ -39,8 +39,8 @@ const console = require('internal/console/global');
 
 const { 0: InspectClient, 1: createRepl } =
     [
-      require('internal/inspector/inspect_client'),
-      require('internal/inspector/inspect_repl'),
+      require('internal/debugger/inspect_client'),
+      require('internal/debugger/inspect_repl'),
     ];
 
 const debuglog = util.debuglog('inspect');
