+++ lib/internal/main/inspect.js
@@ -13,5 +13,5 @@ markBootstrapComplete();
 
 // Start the debugger agent.
 process.nextTick(() => {
-  require('internal/inspector/_inspect').start();
+  require('internal/debugger/_inspect').start();
 });
