+++ node.gyp
@@ -143,6 +143,9 @@
       'lib/internal/crypto/webcrypto.js',
       'lib/internal/crypto/x509.js',
       'lib/internal/constants.js',
+      'lib/internal/debugger/_inspect.js',
+      'lib/internal/debugger/inspect_client.js',
+      'lib/internal/debugger/inspect_repl.js',
       'lib/internal/dgram.js',
       'lib/internal/dns/promises.js',
       'lib/internal/dns/utils.js',
@@ -166,9 +169,6 @@
       'lib/internal/heap_utils.js',
       'lib/internal/histogram.js',
       'lib/internal/idna.js',
-      'lib/internal/inspector/_inspect.js',
-      'lib/internal/inspector/inspect_client.js',
-      'lib/internal/inspector/inspect_repl.js',
       'lib/internal/inspector_async_hook.js',
       'lib/internal/js_stream_socket.js',
       'lib/internal/legacy/processbinding.js',
